Fix engine wiring, repair never-compiled subsystems, harden MCP server
An audit found that much of this engine compiled but never ran, and that a
large part of it had never compiled at all. This connects the existing code to
a running frame and repairs the modules that had drifted out of the build.

Runtime wiring
- Scene::OnUpdate only ticked the UI system, so physics, animation, cameras,
  transforms, and render collection were dead code. Add ECS::SystemPipeline,
  which owns the frame order: input -> character -> physics -> animation -> IK
  -> navigation -> cameras -> world streaming -> transform resolve -> lighting
  and render collection. Includes pause/step/time-scale/fixed-delta control.
- JobSystem::Initialize() was never called anywhere, so the first parallel
  system would have spun forever in Wait(). The pipeline starts and stops it,
  and Execute/Dispatch now run inline when no workers exist instead of hanging.
- Implement RHI::RHIDevice (Core/RHI/Vulkan/VulkanDevice). It had no
  implementation in the tree, which is why TerrainSystem, FoliageSystem and
  SkyboxSystem could not be constructed. Buffers, textures and samplers are
  real; command-list/pipeline/render-pass creation warn once and return null
  because the engine still submits through VulkanContext directly.
- Mesh::vertexBuffer/indexBuffer were declared but never written by any code
  path. Add Mesh::UploadToGPU(RHIDevice&).
- PostProcessManager never instantiated its passes, so the chain was empty at
  runtime. Register SSAO, DoF, motion blur, bloom and color grading in order.

Memory safety and correctness
- AssetLoader: uint32 + uint32 bounds checks wrapped at 2^32, allowing an
  out-of-bounds read from a crafted asset. Widen to 64-bit before comparing.
- AssetLoader: Spirv.resize(size/4) followed by memcpy(size) overflowed the
  heap when a SPIR-V blob was not 4-byte aligned. Reject unaligned sizes.
- EngineJobSystemAdapter: job reference count underflowed (leaking every Jolt
  job) and the hand-rolled barrier ignored SetBarrier failure, so Wait() could
  hang. Replace with Jolt's own JobSystemWithBarrier.
- JobSystem::Shutdown stored isRunning outside the mutex, losing wakeups.
- LuaEngine::ExecuteWithTimeout ignored its timeout parameter, so a script
  reachable from the MCP ExecuteScript tool could run unbounded.
- ClothComponent negated an unsigned before a float multiply, wrapping instead
  of producing a small negative offset.
- FoliageScatterUniforms and AtmosphereUniforms were 16 bytes short of their
  own static_assert.
- AudioSystem::PlaySound linked as PlaySoundA: Windows <mmsystem.h>, reached
  through miniaudio, defines PlaySound as a macro.
- PhysicsSystem called OptimizeBroadPhase every step; only do it after bodies
  are added.

Build
- 18 .cpp files were absent from CMakeLists. Nine of them did not compile at
  all; the exclusion had hidden years of API drift against both their own
  headers and the installed Recast/Detour. All are now fixed and building, and
  no .cpp under Core/ is left out of the build.
- Application.cpp exceeded the COFF section limit from inlining every MCP tool
  family. Move the factory to MCPAllTools.cpp behind a declaration-only header.

MCP server
- CreateAllMCPTools() was never called, so the runtime server exposed zero
  tools, and no tool family compiled. Ten families now build and register.
- tools/call ran on HTTP worker threads, racing the game loop's ECS registry.
  Marshal it onto the main thread with a timeout.
- JSON-RPC notifications were rejected as malformed, breaking the handshake
  with every spec-compliant client. Support notifications and batches.
- Replace wildcard CORS with an Origin allowlist, add optional bearer auth and
  a body size cap.
- Add tool annotations (readOnly/destructive/idempotent/openWorld) and enforce
  the 64-character name limit at registration.
- Add MCPDevTools (engine status, simulation control, log tail, play-mode and
  performance suites, profiler capture) and MCPProjectTools (platform build,
  scene save/load, project listing) with a project-root path sandbox.
- Physics and inventory tools reported success while doing nothing; they now
  return isError with an explicit NOT APPLIED message.
- Add a dependency-free stdio bridge and .mcp.json so stdio MCP clients can
  connect.

Docs and tests
- Add docs/SOTA_GAP_ANALYSIS.md: first-hand comparison against Unreal 5.7 and
  Unity 6 with a dependency-ordered plan. Parity is not reached; asset import
  is the head of the remaining chain.
- Add EngineCoreSystemPipelineTests. Note that ctest runs -C Release, where
  NDEBUG makes assert() a no-op, so the test uses an always-evaluated CHECK.

Verified: ALL_BUILD Release exits 0 with no warnings under /W4 /WX; ctest 19/19;
engine boots headless with 55 MCP tools served and degrades cleanly when no RHI
device is present.

+++ Core/ECS/Components/ClothComponent.h
@@ -196,9 +196,12 @@ namespace ECS {
             for (uint32_t y = 0; y < rows; ++y) {
                 for (uint32_t x = 0; x < cols; ++x) {
                     uint32_t idx = y * cols + x;
+                    // Convert before negating: -y on an unsigned wraps to a huge
+                    // value, which then becomes a huge float instead of a small
+                    // negative offset.
                     Positions[idx] = Math::Vec3(
-                        x * dx - Width * 0.5f,      // Center horizontally
-                        -y * dy + Height * 0.5f,    // Y goes down, center vertically
+                        static_cast<float>(x) * dx - Width * 0.5f,      // Center horizontally
+                        -(static_cast<float>(y) * dy) + Height * 0.5f,  // Y goes down, center vertically
                         0.0f
                     );
                     PreviousPositions[idx] = Positions[idx];
